+++ hub20/apps/ethereum/factories/checkout.py
@@ -2,7 +2,7 @@ import factory
 
 from hub20.apps.core.factories.checkout import CheckoutFactory
 
-from .payments import Erc20TokenPaymentOrderFactory
+from .payments import Erc20TokenPaymentOrderFactory, EtherPaymentOrderFactory
 
 
 class Erc20TokenCheckoutFactory(CheckoutFactory):
@@ -13,4 +13,12 @@ class Erc20TokenCheckoutFactory(CheckoutFactory):
     )
 
 
-__all__ = ["Erc20TokenCheckoutFactory"]
+class EtherCheckoutFactory(CheckoutFactory):
+    order = factory.SubFactory(
+        EtherPaymentOrderFactory,
+        user=factory.SelfAttribute("..store.owner"),
+        reference=factory.fuzzy.FuzzyText(length=30, prefix="checkout-"),
+    )
+
+
+__all__ = ["Erc20TokenCheckoutFactory", "EtherCheckoutFactory"]